app/bot/requirements.txt
loguru
aiogram
python-environ==0.4.54
